:recycle: Organize scripts and ES .json files
