+++ scripts/repairEs.js
@@ -1,5 +1,5 @@
-const esUtils = require('./esUtils');
-const { republishModels } = require('./republishUtils');
+const esUtils = require('./utils/esUtils');
+const { republishModels } = require('./utils/republishUtils');
 const run = async () => {
   await esUtils.deleteSearchIndex();
   await esUtils.createSearchIndex();
