+++ scripts/republish.js
@@ -1,4 +1,4 @@
-const { republishModels } = require('./republishUtils');
+const { republishModels } = require('./utils/republishUtils');
 
 const run = async () => {
   await republishModels();
